+++ source/common/upstream/cds_api_impl.cc
@@ -20,9 +20,10 @@ CdsApiPtr CdsApiImpl::create(const envoy::api::v2::core::ConfigSource& cds_confi
                              ClusterManager& cm, Event::Dispatcher& dispatcher,
                              Runtime::RandomGenerator& random,
                              const LocalInfo::LocalInfo& local_info, Stats::Scope& scope,
+                             ProtobufMessage::ValidationVisitor& validation_visitor,
                              Api::Api& api) {
-  return CdsApiPtr{
-      new CdsApiImpl(cds_config, is_delta, cm, dispatcher, random, local_info, scope, api)};
+  return CdsApiPtr{new CdsApiImpl(cds_config, is_delta, cm, dispatcher, random, local_info, scope,
+                                  validation_visitor, api)};
 }
 
 // TODO(fredlas) the is_delta argument can be removed upon delta+SotW ADS Envoy code unification. It
@@ -33,8 +34,10 @@ CdsApiPtr CdsApiImpl::create(const envoy::api::v2::core::ConfigSource& cds_confi
 CdsApiImpl::CdsApiImpl(const envoy::api::v2::core::ConfigSource& cds_config, bool is_delta,
                        ClusterManager& cm, Event::Dispatcher& dispatcher,
                        Runtime::RandomGenerator& random, const LocalInfo::LocalInfo& local_info,
-                       Stats::Scope& scope, Api::Api& api)
-    : cm_(cm), scope_(scope.createScope("cluster_manager.cds.")) {
+                       Stats::Scope& scope, ProtobufMessage::ValidationVisitor& validation_visitor,
+                       Api::Api& api)
+    : cm_(cm), scope_(scope.createScope("cluster_manager.cds.")),
+      validation_visitor_(validation_visitor) {
   Config::Utility::checkLocalInfo("cds", local_info);
 
   const std::string grpc_method = is_delta ? "envoy.api.v2.ClusterDiscoveryService.DeltaClusters"
@@ -42,8 +45,8 @@ CdsApiImpl::CdsApiImpl(const envoy::api::v2::core::ConfigSource& cds_config, boo
   subscription_ = Config::SubscriptionFactory::subscriptionFromConfigSource(
       cds_config, local_info, dispatcher, cm, random, *scope_,
       "envoy.api.v2.ClusterDiscoveryService.FetchClusters", grpc_method,
-      Grpc::Common::typeUrl(envoy::api::v2::Cluster().GetDescriptor()->full_name()), api, *this,
-      is_delta);
+      Grpc::Common::typeUrl(envoy::api::v2::Cluster().GetDescriptor()->full_name()),
+      validation_visitor, api, *this, is_delta);
 }
 
 void CdsApiImpl::onConfigUpdate(const Protobuf::RepeatedPtrField<ProtobufWkt::Any>& resources,
@@ -51,7 +54,8 @@ void CdsApiImpl::onConfigUpdate(const Protobuf::RepeatedPtrField<ProtobufWkt::An
   ClusterManager::ClusterInfoMap clusters_to_remove = cm_.clusters();
   std::vector<envoy::api::v2::Cluster> clusters;
   for (const auto& cluster_blob : resources) {
-    clusters.push_back(MessageUtil::anyConvert<envoy::api::v2::Cluster>(cluster_blob));
+    clusters.push_back(
+        MessageUtil::anyConvert<envoy::api::v2::Cluster>(cluster_blob, validation_visitor_));
     clusters_to_remove.erase(clusters.back().name());
   }
   Protobuf::RepeatedPtrField<std::string> to_remove_repeated;
@@ -82,7 +86,8 @@ void CdsApiImpl::onConfigUpdate(
   for (const auto& resource : added_resources) {
     envoy::api::v2::Cluster cluster;
     try {
-      cluster = MessageUtil::anyConvert<envoy::api::v2::Cluster>(resource.resource());
+      cluster = MessageUtil::anyConvert<envoy::api::v2::Cluster>(resource.resource(),
+                                                                 validation_visitor_);
       MessageUtil::validate(cluster);
       if (!cluster_names.insert(cluster.name()).second) {
         // NOTE: at this point, the first of these duplicates has already been successfully applied.