+++ test/common/config/utility_test.cc
@@ -27,34 +27,6 @@ using testing::ReturnRef;
 namespace Envoy {
 namespace Config {
 
-TEST(UtilityTest, GetTypedResources) {
-  envoy::api::v2::DiscoveryResponse response;
-  EXPECT_EQ(0, Utility::getTypedResources<envoy::api::v2::ClusterLoadAssignment>(response).size());
-
-  envoy::api::v2::ClusterLoadAssignment load_assignment_0;
-  load_assignment_0.set_cluster_name("0");
-  response.add_resources()->PackFrom(load_assignment_0);
-  envoy::api::v2::ClusterLoadAssignment load_assignment_1;
-  load_assignment_1.set_cluster_name("1");
-  response.add_resources()->PackFrom(load_assignment_1);
-
-  auto typed_resources =
-      Utility::getTypedResources<envoy::api::v2::ClusterLoadAssignment>(response);
-  EXPECT_EQ(2, typed_resources.size());
-  EXPECT_EQ("0", typed_resources[0].cluster_name());
-  EXPECT_EQ("1", typed_resources[1].cluster_name());
-}
-
-TEST(UtilityTest, GetTypedResourcesWrongType) {
-  envoy::api::v2::DiscoveryResponse response;
-  envoy::api::v2::ClusterLoadAssignment load_assignment_0;
-  load_assignment_0.set_cluster_name("0");
-  response.add_resources()->PackFrom(load_assignment_0);
-
-  EXPECT_THROW_WITH_REGEX(Utility::getTypedResources<envoy::api::v2::Listener>(response),
-                          EnvoyException, "Unable to unpack .*");
-}
-
 TEST(UtilityTest, ComputeHashedVersion) {
   EXPECT_EQ("hash_2e1472b57af294d1", Utility::computeHashedVersion("{}").first);
   EXPECT_EQ("hash_33bf00a859c4ba3f", Utility::computeHashedVersion("foo").first);