+++ test/common/stats/isolated_store_impl_test.cc
@@ -45,8 +45,8 @@ TEST_F(StatsIsolatedStoreImplTest, All) {
   c1.add(100);
   EXPECT_EQ(200, found_counter->get().value());
 
-  Gauge& g1 = store_.gauge("g1");
-  Gauge& g2 = scope1->gauge("g2");
+  Gauge& g1 = store_.gauge("g1", Gauge::ImportMode::Accumulate);
+  Gauge& g2 = scope1->gauge("g2", Gauge::ImportMode::Accumulate);
   EXPECT_EQ("g1", g1.name());
   EXPECT_EQ("scope1.g2", g2.name());
   EXPECT_EQ("g1", g1.tagExtractedName());
@@ -108,8 +108,8 @@ TEST_F(StatsIsolatedStoreImplTest, AllWithSymbolTable) {
   EXPECT_EQ(0, c1.tags().size());
   EXPECT_EQ(0, c1.tags().size());
 
-  Gauge& g1 = store_.gaugeFromStatName(makeStatName("g1"));
-  Gauge& g2 = scope1->gaugeFromStatName(makeStatName("g2"));
+  Gauge& g1 = store_.gaugeFromStatName(makeStatName("g1"), Gauge::ImportMode::Accumulate);
+  Gauge& g2 = scope1->gaugeFromStatName(makeStatName("g2"), Gauge::ImportMode::Accumulate);
   EXPECT_EQ("g1", g1.name());
   EXPECT_EQ("scope1.g2", g2.name());
   EXPECT_EQ("g1", g1.tagExtractedName());
@@ -159,12 +159,10 @@ TEST_F(StatsIsolatedStoreImplTest, LongStatName) {
 /**
  * Test stats macros. @see stats_macros.h
  */
-// clang-format off
 #define ALL_TEST_STATS(COUNTER, GAUGE, HISTOGRAM)                                                  \
-  COUNTER  (test_counter)                                                                          \
-  GAUGE    (test_gauge)                                                                            \
+  COUNTER(test_counter)                                                                            \
+  GAUGE(test_gauge, Accumulate)                                                                    \
   HISTOGRAM(test_histogram)
-// clang-format on
 
 struct TestStats {
   ALL_TEST_STATS(GENERATE_COUNTER_STRUCT, GENERATE_GAUGE_STRUCT, GENERATE_HISTOGRAM_STRUCT)
@@ -189,8 +187,7 @@ TEST_F(StatsIsolatedStoreImplTest, NullImplCoverage) {
   NullCounterImpl c(store_.symbolTable());
   EXPECT_EQ(0, c.latch());
   NullGaugeImpl g(store_.symbolTable());
-  g.setShouldImport(true);
-  EXPECT_EQ(absl::nullopt, g.cachedShouldImport());
+  EXPECT_EQ(0, g.value());
 }
 
 } // namespace Stats