+++ test/extensions/filters/common/lua/wrappers_test.cc
@@ -25,7 +25,7 @@ public:
 
   envoy::api::v2::core::Metadata parseMetadataFromYaml(const std::string& yaml_string) {
     envoy::api::v2::core::Metadata metadata;
-    MessageUtil::loadFromYaml(yaml_string, metadata);
+    TestUtility::loadFromYaml(yaml_string, metadata);
     return metadata;
   }
 };